BAB-291: removed transaction base service and moved create logic into transaction service. Transaction in bulk now inherits from transaction service.
